+++ js/data.js
@@ -1,0 +1,45 @@
+export const ARRAY_MAX_COUNT = 25;
+
+/**
+ * @typedef {[min: number, max: number]} NumberRange
+ */
+
+/**
+ * @type {NumberRange}
+ */
+export const COMMENTS_RANGE = [1, 25];
+
+/**
+ * @type {NumberRange}
+ */
+export const LIKES_RANGE = [15, 200];
+
+/**
+ * @type {NumberRange}
+ */
+export const AVATAR_RANGE = [1, 6];
+
+export const DESCRIPTIONS = [
+  'описание 1',
+  'описание 2',
+  'описание 3',
+  'описание 4',
+];
+
+export const MESSAGES = [
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Как можно было поймать такой неудачный момент?!',
+];
+
+export const NAMES = [
+  'Иван',
+  'Хуан Себастьян',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+];
